splitting instruments test
